exshape v0.1.0 Exshape

This module just contains a helper function for working wtih zip archives. If you have a stream of bytes that you want to parse directly, use the Shp or Dbf modules to parse.

Types

layer()

layer() :: {layer_name, {projection, Stream.t}}

layer_name()

layer_name() :: String.t

projection()

projection() :: String.t

Functions

from_zip(path, opts \\ [])

Given a zip file path, unzip it and open streams for the underlying shape data.

Returns a list of all the layers, where each layer is a tuple of projection and the stream of features